+++ src/js/_init.js
@@ -1,4 +1,38 @@
 (function ($) {
+    EqUI.init = {};
+    var _this = function(){return EqUI.init;}();
+
+    // Update
+    _this.update = function() {
+        EqUI.app_bar.update();
+        EqUI.side_nav.update();
+        EqUI.breadcrumb.update();
+
+        EqUI.buttons.update();
+        EqUI.cards.update();
+        EqUI.forms.update();
+        EqUI.collapsible.update();
+        EqUI.dropdown.update();
+        EqUI.modals.update();
+        EqUI.tabs.update();
+
+        EqUI.site.update();
+    };
+
+    // Load
+    _this.load = function() {
+        // Site update
+        EqUI.site.update();
+
+        // Show body
+        EqUI.site.body.css('visibility', 'visible');
+
+        // Loads
+        EqUI.breadcrumb.load();
+        EqUI.dropdown.load();
+        EqUI.tabs.load();
+    };
+
 	$(document).ready(function() {
 
 		// Init Waves
@@ -8,33 +42,14 @@
 
 		// Resize
 		$(window).resize( function() {
-			
-			EqUI.app_bar.update();
-			EqUI.side_nav.update();
-
-			EqUI.buttons.update();
-			EqUI.cards.update();
-			EqUI.forms.update();
-			EqUI.collapsible.update();
-            EqUI.dropdown.update();
-			EqUI.modals.update();
-			EqUI.tabs.update();
-			
-			EqUI.site.update();
-
+            // Update
+            _this.update();
 		});
 
 		// Load complete
-		$(window).load(function(){ 
-	        // Site update
-	        EqUI.site.update();
-
-	        // Show body
-            EqUI.site.body.css('visibility', 'visible');
-
-            // Loads
-            EqUI.dropdown.load();
-            EqUI.tabs.load();
+		$(window).load(function(){
+            // Update
+            _this.load();
 	    });
 	    
 	});